requirements.txt
ultralytics==8.3.0
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.2
streamlit==1.39.0
scikit-image==0.24.0
